Add maintenance views for managing plans and reports, including creation and editing functionality
